Assorted UI cleanups. Add a bunch of little image files that may be used in the Ouya port later.
