+++ UI/MiscScreens.cpp
@@ -1,0 +1,64 @@
+// Copyright (c) 2013- PPSSPP Project.
+
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, version 2.0 or later versions.
+
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License 2.0 for more details.
+
+// A copy of the GPL 2.0 should have been included with the program.
+// If not, see http://www.gnu.org/licenses/
+
+// Official git repository and contact information can be found at
+// https://github.com/hrydgard/ppsspp and http://www.ppsspp.org/.
+
+#include "base/colorutil.h"
+#include "base/timeutil.h"
+#include "gfx_es2/draw_buffer.h"
+#include "ui/ui_context.h"
+#include "ui/view.h"
+#include "ui/viewgroup.h"
+#include "UI/MiscScreens.h"
+
+void DrawBackground(float alpha);
+
+void PromptScreen::DrawBackground(UIContext &dc) {
+	::DrawBackground(1.0f);
+}
+
+void PromptScreen::CreateViews() {
+	// Information in the top left.
+	// Back button to the bottom left.
+	// Scrolling action menu to the right.
+	using namespace UI;
+
+	Margins actionMenuMargins(0, 100, 15, 0);
+
+	root_ = new LinearLayout(ORIENT_HORIZONTAL);
+
+	ViewGroup *leftColumn = new AnchorLayout(new LinearLayoutParams(1.0f));
+	root_->Add(leftColumn);
+
+	leftColumn->Add(new TextView(0, message_, ALIGN_LEFT, 1.0f, new AnchorLayoutParams(10, 10, NONE, NONE)));
+
+	ViewGroup *rightColumnItems = new LinearLayout(ORIENT_VERTICAL, new LinearLayoutParams(300, FILL_PARENT, actionMenuMargins));
+	root_->Add(rightColumnItems);
+	rightColumnItems->Add(new Choice(yesButtonText_))->OnClick.Handle(this, &PromptScreen::OnYes);
+	if (noButtonText_ != "")
+		rightColumnItems->Add(new Choice(noButtonText_))->OnClick.Handle(this, &PromptScreen::OnNo);
+}
+
+UI::EventReturn PromptScreen::OnYes(UI::EventParams &e) {
+	callback_(true);
+	screenManager()->finishDialog(this, DR_OK);
+	return UI::EVENT_DONE;
+}
+
+UI::EventReturn PromptScreen::OnNo(UI::EventParams &e) {
+	callback_(false);
+	screenManager()->finishDialog(this, DR_CANCEL);
+	return UI::EVENT_DONE;
+}
