+++ Windows/WndMainWindow.cpp
@@ -896,13 +896,15 @@ namespace MainWindow
 					if (raw->data.keyboard.Message == WM_KEYDOWN || raw->data.keyboard.Message == WM_SYSKEYDOWN) {
 						key.flags = KEY_DOWN;
 						key.keyCode = windowsTransTable[GetTrueVKey(raw->data.keyboard)];
-						if (key.keyCode)
+						if (key.keyCode) {
 							NativeKey(key);
+						}
 					} else if (raw->data.keyboard.Message == WM_KEYUP) {
 						key.flags = KEY_UP;
 						key.keyCode = windowsTransTable[GetTrueVKey(raw->data.keyboard)];
-						if (key.keyCode)
+						if (key.keyCode) {
 							NativeKey(key);	
+						}
 					}
 				}
 			}
